echo, envの追加とbuiltinコマンドの移動
