+++ src_resaito/builtin/echo.c
@@ -1,32 +1,38 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_env.c                                           :+:      :+:    :+:   */
+/*   echo.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: resaito <resaito@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/12/08 15:24:50 by resaito           #+#    #+#             */
-/*   Updated: 2023/12/08 15:55:22 by resaito          ###   ########.fr       */
+/*   Created: 2023/12/08 16:07:03 by resaito           #+#    #+#             */
+/*   Updated: 2023/12/08 16:52:48 by resaito          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
+#include "../../inc/minishell.h"
 
-int	ft_env(t_envval *envval)
+int	echo(t_node *node)
 {
-	char	**str;
+	bool	option;
 	size_t	len;
 
-	if (!(envval->env))
-		return (1);
-	str = make_env_strs(envval->env);
-	len = 0;
-	while (str[len])
+	option = false;
+	len = 1;
+	if (node->args[1] && ft_strcmp(node->args[1], "-n") == 0)
 	{
-		printf("%s\n", str[len]);
+		option = true;
+		len = 2;
+	}
+	while (node->args[len])
+	{
+		ft_putstr_fd(node->args[len], STDOUT_FILENO);
+		if (node->args[len + 1])
+			write(STDOUT_FILENO, " ", 1);
 		len++;
 	}
-    str_array_free(str);
+	if (!option)
+		write(STDOUT_FILENO, "\n", 1);
 	return (0);
 }
 
@@ -43,19 +49,14 @@ int	ft_env(t_envval *envval)
 
 // int	main(int ac, char **av, char **envp)
 // {
-// 	// t_node		*node;
-// 	t_envval	*envval;
-// 	// char		*env_arg[] = {"env", NULL};
+// 	t_node		*node;
+// 	char		*echo_arg[] = {"echo", "aaaa", "aaaa", "aaaa", "aaaa",
+// 				NULL};
 
-// 	// node = make_node(N_COMMAND, env_arg);
-// 	envval = make_envval(new_envs(envp)); //make_envval.c make_env.c
-// 	ft_env(envval);
-// 	// free(node);
-// 	envs_free(envval->env); //env_free.c ../src/free.c
-// 	free(envval);
+// 	node = make_node(N_COMMAND, echo_arg);
+// 	// envval = make_envval(new_envs(envp)); //make_envval.c make_env.c
+// 	echo(node);
+// 	free(node);
+// 	// envs_free(envval->env); //env_free.c ../src/free.c
+// 	// free(envval);
 // } // ../libft/libft.a
-
-// __attribute__((destructor))
-// static void destructor() {
-// 	system("leaks -q a.out");
-// }
