+++ src_resaito/builtin/pwd.c
@@ -6,11 +6,11 @@
 /*   By: resaito <resaito@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/08 13:26:22 by resaito           #+#    #+#             */
-/*   Updated: 2023/12/08 13:26:42 by resaito          ###   ########.fr       */
+/*   Updated: 2023/12/08 16:08:55 by resaito          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
+#include "../../inc/minishell.h"
 
 int	pwd(t_envval *envval)
 {
